Use tabs in an app bar for the whole page
