+++ src/components/Options/Custom.js
@@ -10,6 +10,7 @@ export default function Custom({ custom, onCustom }) {
       onChange={onCustom}
       type="number"
       min="0"
+      fullWidth
     />
   )
 }
