+++ src/components/Options/index.test.js
@@ -4,12 +4,10 @@ import React from 'react'
 import Options from '.'
 
 test('Options', () => {
-  const handleOptionsKey = jest.fn()
   const wrapper = shallow(
     <Options
       onCustom={() => {}}
       onGeneric={() => {}}
-      onOptionsKey={handleOptionsKey}
       options={{
         Custom: 0,
         Generic: {
@@ -22,7 +20,5 @@ test('Options', () => {
       optionsKey="Generic"
     />
   )
-  wrapper.find('WithStyles(Tabs)').simulate('change')
   expect(wrapper).toMatchSnapshot()
-  expect(handleOptionsKey).toHaveBeenCalled()
 })
